requirements.txt
gymnasium==0.29.1
pygame==2.6.0
torch==2.4.1
numpy==1.26.4